+++ Credixware/Internal/Hacks/Skinchanger.h
@@ -6,8 +6,16 @@
 
 /* NOTE: Changed by FireEventClientSide.h */
 bool bKnifeEquipped = false;
+int iLocalID = 0;
 
 namespace Skinchanger {
+	void Init() {
+		player_info_t localPlayer;
+
+		if (g_pEngineClient->GetPlayerInfo(g_pEngineClient->GetLocalPlayer(), &localPlayer)) {
+			iLocalID = localPlayer.user_id;
+		}
+	}
 	void Update() {
 		/*IClientEntity* LocalPlayer = g_pClientEntityList->GetClientEntity(g_pEngineClient->GetLocalPlayer());
 		UINT* hWeapons = (UINT*)((DWORD)LocalPlayer->GetBaseEntity() + 0x2DF8);
@@ -19,7 +27,7 @@ namespace Skinchanger {
 			IClientEntity* Entity = g_pClientEntityList->GetClientEntity(i);
 			if (Entity) {
 				ClientClass* EntityClass = Entity->GetClientClass();
-				if (EntityClass->m_ClassID == 136 /* || !_strcmpi("CPredictedViewModel", EntityType) */) {
+				if (EntityClass->m_ClassID == 136 /* same as "_strcmpi("CPredictedViewModel", EntityType) == 0" */) {
 					C_BaseEntity* BaseEntity = Entity->GetBaseEntity();
 					if (BaseEntity && bKnifeEquipped) {
 						int karam = g_pModelInfo->GetModelIndex("models/weapons/v_knife_karam.mdl");