+++ Credixware/Internal/Headers/Offsets.h
@@ -1,0 +1,12 @@
+// Created by Credix - Licence: https://creativecommons.org/licenses/by-nc-sa/4.0/ 
+#ifndef OFFSETS_H
+#define OFFSETS_H
+
+#include "Common.h"
+
+namespace Offsets {
+	DWORD m_fFlags = 0x104;
+	DWORD m_iHealth = 0x100;
+}
+
+#endif